refactor 2x2
