--- src/app/[lang]/indian/layout.tsx
@@ -1,7 +1,0 @@
-export default function TwoDigitLayout({
-    children,
-}: Readonly<{
-    children: React.ReactNode
-}>) {
-    return <>{children}</>
-}
