+++ src/app/[lang]/indian/questionBox.module.scss
@@ -4,7 +4,6 @@
     justify-content: center;
     align-items: center;
 }
-
 .question {
     padding: 16px;
     font-size: 4rem;
