+++ src/app/[lang]/two-digit-x-two-digit/createNumbers.ts
@@ -1,0 +1,7 @@
+export function createNumbers(): number[] {
+    const firstNumber = Math.floor(Math.random() * 89) + 11
+
+    const secondNumber = Math.floor(Math.random() * 89) + 11
+
+    return [firstNumber, secondNumber]
+}
