--- src/app/[lang]/two-digit-x-two-digit/layout.tsx
@@ -1,7 +1,0 @@
-export default function TwoDigitLayout({
-    children,
-}: Readonly<{
-    children: React.ReactNode
-}>) {
-    return <>{children}</>
-}
