+++ src/app/[lang]/two-digit-x-two-digit/questionBox.module.scss
@@ -1,0 +1,10 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+.question {
+    padding: 16px;
+    font-size: 4rem;
+}
